backend/requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohttp==3.9.5
aiokafka==0.10.0
aioredis==1.3.1
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.2.0
APScheduler==3.10.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
azure-core==1.30.1
azure-eventhub==5.12.0
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.20.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
build==1.2.1
certifi==2024.2.2
cffi==1.16.0
channels==3.0.4
channels-redis==3.2.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.7
cssselect2==0.7.0
daphne==3.0.2
databases==0.9.0
dataclasses==0.6
distro==1.9.0
dj-database-url==0.5.0
Django==5.0.7
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap-daterangepicker==1.1.0
django-clearcache==1.2.1
django-cors-headers==3.7.0
django-filter==24.2
django-location-field==2.7.3
django-xhtml2pdf==0.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.5.0
drf-yasg==1.21.7
email_validator==2.2.0
et-xmlfile==1.1.0
Faker==22.6.0
fastapi==0.111.1
fastapi-cli==0.0.4
Flask==3.0.3
Flask-Cors==4.0.1
flask-restplus==0.13.0
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
gast==0.5.5
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
hiredis==3.0.0
html5lib==1.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
keras==3.4.1
libclang==18.1.1
line_profiler==4.1.3
loguru==0.7.2
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
motor==3.3.2
msal==1.28.0
msal-extensions==1.1.0
msgpack==1.0.8
msrest==0.7.1
multidict==6.0.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
openai==1.33.0
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
oscrypto==1.3.0
packaging==24.0
pandas==2.2.2
pdfkit==1.0.0
peewee==3.17.0
Pillow==8.3.1
pip-tools==7.4.1
portalocker==2.8.2
protobuf==4.25.3
psutil==5.9.8
psycopg==3.2.1
psycopg2-binary==2.9.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycares==4.4.0
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.1.0
pydantic_core==2.18.2
pydyf==0.10.0
Pygments==2.18.0
pyHanko==0.23.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pymongo==4.6.1
PyMySQL==1.1.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pypdf==4.1.0
pyphen==0.15.0
pypng==0.20220715.0
pyproject_hooks==1.1.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.35.1
regex==2024.5.15
reportlab==4.0.9
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
schedule==1.2.1
service-identity==24.1.0
setuptools==70.1.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.0
starlette==0.37.2
svglib==1.5.1
swagger_ui==0.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.12.0
tensorflowjs==3.18.0
termcolor==2.4.0
tiktoken==0.7.0
tinycss2==1.3.0
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.2.1
uuid==1.30
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
weasyprint==62.2
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.5.0
wonderwords==2.2.0
wrapt==1.16.0
xhtml2pdf==0.2.15
yappi==1.6.0
yarl==1.9.4
zope.interface==6.4.post2
zopfli==0.2.3
